refactor (utils): implement enum utils
